Move all sample python files into python folder
