+++ src/main/resources/sample/python/ChemEntity.py
