+++ src/main/resources/sample/python/ChemTypes.py
