+++ src/main/resources/sample/python/Equality.py
